chore: refactoring of the feedback module to make it domain-specific & thanks to it prevent the circular dependency
