+++ suite-common/feedback/src/index.ts
@@ -1,4 +1,4 @@
-export * from './feedbackActions';
-export * from './feedbackUtils';
-export * from './types';
-export * from './constants';
+export * from './sendFeedbackAction';
+export * from './userData';
+export * from './rating';
+export * from './feedback';
